Changes after Romain's Issue : created json files instead of var in main, sorted functions, general clean and start to think about props/states
