--- data/test.json
@@ -1,4 +1,0 @@
-{
-    part1: 1,
-    part2: 2
-}
