+++ src/Components/BooleanFilter.js
@@ -1,12 +1,10 @@
-
-
 'use strict';
 
 var React = require('react');
     
 /*
     props: {
-        active: boolean,
+        fav: boolean,
         className: String,
         onChange: (nextStat: boolean) => void
     }
@@ -26,20 +24,7 @@ module.exports = React.createClass({
         
         return React.DOM.button({className: [props.className, (state.active ? 'active' : '')].join(' '),
             onClick: function(){
-                var nextState = !state.active;
-                self.setState({active: nextState});
-                props.onChange(nextState);
+                props.onChange(!propS.fav);
             }});
-                                
-        
-        
-/*        return React.DOM.li({className: [props.className, (state.active ? 'active' : '')].join(' ')}, 
-            React.DOM.button({
-                onClick: function(){
-                    var nextState = !state.active;
-                    self.setState({active: nextState});
-                    props.onChange(nextState);
-                }
-            }, props.label));*/
     }
 });
