+++ src/dateLists.js
@@ -1,0 +1,9 @@
+'use strict';
+
+var weekDays = ['lundi', 'mardi', 'mercredi', 'jeudi' , 'vendredi', 'samedi' , 'dimanche'];
+var months = [ "janvier", "février", "mars", "avril", "mai", "juin",
+                  "juillet", "août", "septembre", "octobre", "novembre", "décembre"];
+module.exports = {
+    weekDays : weekDays,
+    months : months
+};
