+++ src/pictoDictionary.json
@@ -1,0 +1,50 @@
+{
+    "Ameublement" : "../pictures/picto_RC/Cartons.svg",
+    "Amiante_Ciment" : "../pictures/picto_RC/Cartons.svg",
+    "Batteries" : "../pictures/picto_RC/Cartons.svg",
+    "Bois" : "../pictures/picto_RC/Bois.svg",
+    "Bois_traite" : "../pictures/picto_RC/Cartons.svg", 
+    "Bouteilles_de_gaz": "../pictures/picto_RC/Cartons.svg",
+    "Bouteilles_Plastiques" : "../pictures/picto_RC/Cartons.svg",
+    "Cartons" : "../pictures/picto_RC/Cartons.svg",
+    "Cartouche_Encre" : "../pictures/picto_RC/Cartons.svg",
+    "Deblais-gravats" : "../pictures/picto_RC/Cartons.svg",
+    "Decheterie_Picto_DDS" : "../pictures/picto_RC/Cartons.svg",
+    "Dechets_d'activites_de_soins_a_risques" : "../pictures/picto_RC/Cartons.svg",
+    "Dechets_verts" : "../pictures/picto_RC/Dechets_verts.svg",
+    "DEEE" : "../pictures/picto_RC/Cartons.svg",
+    "Ecrans" : "../pictures/picto_RC/Cartons.svg",
+    "Encombrants_1" : "../pictures/picto_RC/Cartons.svg",
+    "Encombrants_2" : "../pictures/picto_RC/Cartons.svg",
+    "Extincteurs" : "../pictures/picto_RC/Cartons.svg",
+    "Films_agricoles_usages" : "../pictures/picto_RC/Cartons.svg",
+    "Gros_electromenager_0" : "../pictures/picto_RC/Cartons.svg",
+    "Huiles_de_fritures" : "../pictures/picto_RC/Cartons.svg",
+    "Huiles_de_Vidange" : "../pictures/picto_RC/Cartons.svg",
+    "Journaux_et_Revues" : "../pictures/picto_RC/Cartons.svg",
+    "Lampes" : "../pictures/picto_RC/Cartons.svg",
+    "Liquides_de_freins" : "../pictures/picto_RC/Cartons.svg",
+    "Liquides_de_refroidissement" : "../pictures/picto_RC/Cartons.svg",
+    "LogoDecheterie" : "../pictures/picto_RC/Cartons.svg",
+    "Medicaments" : "../pictures/picto_RC/Cartons.svg",
+    "Metaux" : "../pictures/picto_RC/Cartons.svg",
+    "Mobilier" : "../pictures/picto_RC/Cartons.svg",
+    "Papiers" : "../pictures/picto_RC/Cartons.svg",
+    "Papiers_cartons" : "../pictures/picto_RC/Cartons.svg",
+    "Pelouse" : "../pictures/picto_RC/Cartons.svg",
+    "Petits_appareils_menagers" : "../pictures/picto_RC/Cartons.svg",
+    "Piles_Boutons" : "../pictures/picto_RC/Cartons.svg",
+    "Piles_et_Accumulateurs" : "../pictures/picto_RC/Cartons.svg",
+    "Plastiques" : "../pictures/picto_RC/Cartons.svg",
+    "Platre_et_plaques_de_platre" : "../pictures/picto_RC/Cartons.svg",
+    "Pneumatiques" : "../pictures/picto_RC/Cartons.svg",
+    "Polystyrene" : "../pictures/picto_RC/Cartons.svg",
+    "Radiographies" : "../pictures/picto_RC/Cartons.svg",
+    "Refrigerateurs-Congelateurs" : "../pictures/picto_RC/Cartons.svg",
+    "Reutilisation_reemploi" : "../pictures/picto_RC/Cartons.svg",
+    "Tailles" : "../pictures/picto_RC/Cartons.svg",
+    "Textiles" : "../pictures/picto_RC/Cartons.svg",
+    "Tout_venant_incinerable" : "../pictures/picto_RC/Cartons.svg",
+    "Tout_venant_non_incinerable" : "../pictures/picto_RC/Cartons.svg",
+    "Verres" : "../pictures/picto_RC/Cartons.svg",
+}
