+++ src/userFake.json
@@ -1,0 +1,4 @@
+{
+    
+    favouriteRC : undefined
+}
